Finished python 3 compatibility.

+++ chapter10/datatypes/video.py
@@ -63,7 +63,7 @@ class Video(Base):
 
         data = self.data[:]
 
-        for i in range(len(data) / (188 + (8 if self.iph else 0))):
+        for i in range(int(len(data) / (188 + (8 if self.iph else 0)))):
             if self.iph:
                 self.all.append(Data('IPH', data[:8]))
                 data = data[8:]
